map initial and cuiser component first attempt
