+++ ruoyi-simulation/src/main/resources/front-dev/traffic_demo/public/index.html
@@ -4,13 +4,18 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <script type="text/javascript">
+        window._AMapSecurityConfig = {
+            securityJsCode: "abeb1094acf058e343c316faf2ce00cd",
+        };
+    </script>
     <title>智慧交通大模型</title>
 </head>
 
 <body>
     <div id="root">
-
     </div>
+
     <script type="text/javascript" src="../statics/record.js"></script>
     <script type="text/javascript" src="../statics/webRtcPlayer.js"></script>
     <script type="text/javascript" src="../statics/loadingPixStream.js"></script>
