+++ ruoyi-simulation/src/main/resources/front-dev/traffic_demo/src/App.js
@@ -2,7 +2,8 @@ import React from 'react';
 import Header from 'components/Header';
 import PlanMain from 'components/PlanMain';
 import JunctionMain from 'components/JunctionMain';
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import CruiserMain from 'components/CruiserMain';
+import { HashRouter as Router, Route, Routes } from 'react-router-dom';
 
 export default function App() {
   return (
@@ -12,6 +13,7 @@ export default function App() {
         <Routes>
           <Route path="/" element={<PlanMain />} />
           <Route path="/junction" element={<JunctionMain />} />
+          <Route path="/cruiser" element={<CruiserMain />} />
         </Routes>
       </div>
     </Router>
