+++ ruoyi-simulation/src/main/resources/front-dev/traffic_demo/src/components/Header.js
@@ -8,7 +8,7 @@ export default function Header() {
             <ul className='nav-left'>
                 <li><Link to='/'>路况概览</Link></li>
                 <li><Link to='/Junction'>路口场景</Link></li>
-                <li>车辆跟踪</li>
+                <li><Link to='/Cruiser'>车辆跟踪</Link></li>
             </ul>
             <div className="title">
                 <h2 className="title-1">智慧交通大模型</h2>
