+++ ruoyi-simulation/src/main/resources/front-dev/traffic_demo/src/components/container/Weather/index.scss
@@ -1,6 +1,47 @@
 .weather-time-container {
-    .time {
-        font-size: 1.3rem;
-        color: rgba(67, 204, 107, 0.7);
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
+    width: 20%;
+    position: absolute;
+    top: 0;
+    left: 0;
+    height: 100%;
+    .time-contianer {
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+        align-items: center;
+        .date {
+            display: flex;
+            flex-direction: row;
+            justify-content: space-between;
+            align-items: flex-start;
+            margin: 0 0.5rem;
+            vertical-align: middle;
+            span {
+                vertical-align: middle;
+                margin: 0 0.5rem;
+            }
+        }
+        .time {
+            font-size: 1.2rem;
+            color: rgba(67, 204, 107, 1);
+        }
+    }
+
+    .weather-container {
+        width: 50%;
+        display: flex;
+        justify-content: space-around;
+        align-items: center;
+        .weather {
+            .weatherIcon SVG {
+                $IConSize: 1.5rem;
+                width: $IConSize;
+                height: $IConSize;
+            }
+        }
     }
 }
