--- ruoyi-simulation/src/main/resources/front-dev/traffic_demo/src/components/container/crusier-view/sidebar-left/Bottom/style.scss
@@ -1,94 +1,0 @@
-.leftBottom {
-    font-family: sans-serif; // Use appropriate font
-    background-color: #1e2e4f;
-    width: 450px; // Adjust width as required
-    height: 350px;
-    padding-bottom: 20px;
-    .header {
-        background-color: rgba(28, 42, 71, 0.7);
-        font-size: 1.5rem;
-        height: 3rem;
-        line-height: 3rem;
-        color: white;
-        align-items: center;
-        text-align: center;
-        position: relative;
-        span {
-            text-shadow:
-                1px 1px 1px #fff,
-                -1px -1px 1px #000;
-        }
-        .right-text {
-            font-size: 15px;
-            position: absolute;
-            right: 10px;
-        }
-    }
-    .head {
-        margin: 2rem 0 1rem 0.3rem;
-        font-size: 1.1rem;
-        ul {
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-        }
-    }
-    .list-container {
-        color: #ddd;
-        height: 230px;
-        overflow: auto;
-        &::-webkit-scrollbar {
-            width: 0;
-        }
-        &::-webkit-scrollbar-track {
-            background: transparent;
-        }
-        &::-webkit-scrollbar-thumb {
-            background: transparent;
-        }
-        > div:nth-child(odd) {
-            background-color: rgba(28, 42, 71, 0.5);
-        }
-    }
-    .list-item {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        padding: 20px 0;
-        span {
-            width: 25%;
-            text-align: right;
-        }
-        .rank {
-            width: 20px;
-            text-align: center;
-        }
-
-        .name {
-            text-overflow: ellipsis;
-            white-space: nowrap;
-            overflow: hidden;
-            text-align: left;
-            margin-left: 5rem;
-        }
-
-        .index {
-            text-align: center;
-        }
-
-        .status {
-            padding: 2px 5px;
-            text-align: right;
-        }
-        span.red{
-            color: red;
-        }
-        span.green{
-            color: green;
-        }
-        .grade {
-            width: 20px;
-            text-align: center;
-        }
-    }
-}
