+++ ruoyi-simulation/src/main/resources/front-dev/traffic_demo/src/components/container/junction-view/sidebar-right/index.jsx
@@ -140,7 +140,6 @@ export default function Index() {
             </div>
         )
     });
-    console.log(renderList);
     return (
         <section className="slider-right">
