+++ ruoyi-simulation/src/main/resources/front-dev/traffic_demo/src/components/container/plan-view/bottom/TrafficRank/TrafficRank.jsx
@@ -61,7 +61,7 @@ export default function TrafficRank() {
         )
     })
 
-    console.log(renderList);
+    // console.log(renderList);
 
     return (
         <div className="traffic-rank">
