+++ ruoyi-simulation/src/main/resources/front-dev/traffic_demo/src/legacy/crusier-view/sidebar-left/Top/style.scss
@@ -1,0 +1,45 @@
+.leftTop{
+    width: 450px;
+    height: 150px;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: flex-start;
+
+    .title{
+        font-size: 1.5rem;
+        font-weight: bold;
+        text-align: center;
+        background-color: rgba(28, 42, 71, 0.7);
+        width: 100%;
+        height: 3rem;
+        line-height: 3rem;
+        vertical-align: center;
+    }
+    .statistic{
+        margin: 2rem auto;
+        display: grid;
+        grid-template-columns: 1fr 1fr 1fr;
+        gap: 0.5rem;
+    }
+    .stat-item{
+        display: flex;
+        justify-content: space-around;
+        align-items: center;
+        .label{
+            text-align: center;
+            text-wrap: nowrap;
+            color: #ddd;
+        }
+        .value{
+            margin: 0 0.3rem;
+            font-weight: bold;
+            color: #74d86d;
+        }
+    }
+    .icon{
+        width: 1.5rem;
+        height: 1.5rem;
+    }
+
+}
