+++ ruoyi-simulation/src/main/resources/front-dev/traffic_demo/src/legacy/crusier-view/sidebar-right/Bottom/style.scss
@@ -1,0 +1,24 @@
+.rightBottom {
+    position: absolute !important;
+    top: 450px;
+    width: 450px;
+    height: 350px;
+    margin-bottom: 5rem;;
+    h1{
+        background-color: rgba(28, 42, 71, 0.7);
+        font-size: 1.5rem;
+        height: 3rem;
+        width: 100%;
+        line-height: 3rem;
+        color: white;
+        text-align: center;
+    }
+    div {
+        position: absolute;
+        left: 50%;
+        top: 50%;
+        transform: translate(-50%, -50%);
+        height: 80%;
+        width: 80%;
+    }
+}
