+++ ruoyi-simulation/src/main/resources/front-dev/traffic_demo/src/legacy/crusier-view/sidebar-right/Top/style.scss
@@ -1,0 +1,75 @@
+@mixin stat-detail-layout {
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: space-evenly;
+    height: 50%;
+    .stat-item-date {
+        background-color: #1f2b65;
+        height: 2rem;
+        border: 1px solid #aaa;
+        border-radius: 0.5rem;
+        padding: 0.5rem;
+        margin-top: 0.5rem;
+    }
+    .stat-item-number {
+        font-weight: bold;
+        font-size: 3rem;
+    }
+    .decrease {
+        width: 100%;
+        text-align: center;
+        span {
+            margin: 1rem;
+        }
+    }
+}
+$green: #00ff3d;
+$red: #ff0000;
+
+.rightTop {
+    display: grid;
+    grid-template-rows: 1fr 4fr 4fr;
+    background-color: #1a1a2d;
+    color: #fff;
+    width: 450px;
+    height: 400px;
+    h1 {
+        background-color: rgba(28, 42, 71, 0.7);
+        font-size: 1.5rem;
+        height: 3rem;
+        width: 100%;
+        line-height: 3rem;
+        color: white;
+        text-align: center;
+    }
+    .stats {
+        display: grid;
+        grid-template-columns: 1fr 1fr;
+        .stat-situation {
+            display: flex;
+            flex-direction: column;
+            justify-content: space-between;
+            align-items: center;
+            border-right: 1px dashed #aaa;
+            .stat-current {
+                @include stat-detail-layout;
+            }
+            .stat-today{
+                @include stat-detail-layout;
+            }
+        }
+        #pieChart {
+            height: 100%;
+            width: 100%;
+        }
+    }
+    #lineChart {
+        height: 100%
+    }
+    .green {
+        color: $green;
+    }
+    .red{
+        color: $red;
+    }
+}
